Remove BUNDLED WITH from Gemfile.lock
to maintain compatibility for Bundler 1+2.
This enables the app to be used with Ruby 2.7, as well as with Ruby 2.6 and
earlier.
https://github.com/sclorg/ruby-ex/pull/34#issuecomment-668557635

More context: Ruby 2.7 is shipped with Bundler 2 which requires Gemfile.lock
update (bundle update --bundler) to be able to work with it. That regenerated
Gemfile.lock is howerver not usable with Bundler 1 (shipped with Ruby 2.6 and
earlier) anymore. BUNDLED WITH removal works around this issue.
https://fedoraproject.org/wiki/Changes/Bundler_2.0